trims out unnessicary checks on user data
related #147
